mimasv2: Cleanup and modernize.
